% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{nldr_viz_server}
\alias{nldr_viz_server}
\title{NLDR Visualization Tool Server}
\usage{
nldr_viz_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
A Shiny server function
}
\description{
Creates the server logic for the NLDR visualization tool
}
