% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{four_clusters}
\alias{four_clusters}
\title{Four clusters dataset (internal use)}
\format{
A data frame with 800 rows and 2 variables.
}
\usage{
four_clusters
}
\description{
Four clusters dataset (internal use)
}
\keyword{internal}
