% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{load_custom_datasets}
\alias{load_custom_datasets}
\title{Load Custom Datasets}
\usage{
load_custom_datasets()
}
\value{
A list of datasets
}
\description{
Loads datasets from the package data or uses built-in datasets
}
