% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_nldr_viz}
\alias{run_nldr_viz}
\title{Run NLDR Visualization Tool}
\usage{
run_nldr_viz(...)
}
\arguments{
\item{...}{Additional arguments passed to shiny::shinyApp}
}
\value{
Shiny application object
}
\description{
Launches the NLDR Visualization Tool Shiny application
}
\examples{
\dontrun{
run_nldr_viz()
}
}
